refactor(mistralai_engine): extract OCR functionality into a separate module
- Moved OCR-related methods from `mistralai_engine.py` to a new module `mistralai_ocr.py` to enhance code modularity and maintainability.
- Updated `handle_ocr` in `mistralai_engine.py` to utilize the `handle_ocr` function from the new module.
- Removed unused imports and corrected import paths in `mistralai_engine.py`.
- Fixed a typo in a model description string.
- Cleaned up docstring comments in `conversation_tab.py` for better clarity.
